GW-22: Refactored repo
