+++ src/pages/home/components/LandingImage.js
@@ -2,7 +2,7 @@ import React from "react";
 import { makeStyles } from "@mui/styles";
 import { CustomButton } from "../../../shared/components/CustomButton";
 import Gisau from "../../../assets/gisau-logo/gisau.png";
-import LandingImg from "../../../assets/landing-image/1.JPG";
+import LandingImg from "../../../assets/landing.jpg";
 
 const useStyles = makeStyles({
   bgImage: {
